Default state for page and rowsPerPage + refactor pagination component
