+++ src/Service.Host/client/src/components/salesPackages/SalesPackages.js
@@ -13,11 +13,8 @@ import {
 } from '@material-ui/core';
 import { Link } from 'react-router-dom';
 import EditIcon from '@material-ui/icons/Edit';
-import IconButton from '@material-ui/core/IconButton';
-import FirstPageIcon from '@material-ui/icons/FirstPage';
-import KeyboardArrowLeft from '@material-ui/icons/KeyboardArrowLeft';
-import KeyboardArrowRight from '@material-ui/icons/KeyboardArrowRight';
-import LastPageIcon from '@material-ui/icons/LastPage';
+
+import TablePaginationActions from '../common/TablePaginationActions';
 import { getSelfHref } from '../../helpers/utilities';
 import Page from '../../containers/Page';
 
@@ -29,79 +26,14 @@ const actionsStyles = theme => ({
     }
 });
 
-class TablePaginationActions extends React.Component {
-    handleFirstPageButtonClick = event => {
-        this.props.onChangePage(event, 0);
-    };
-
-    handleBackButtonClick = event => {
-        this.props.onChangePage(event, this.props.page - 1);
-    };
-
-    handleNextButtonClick = event => {
-        this.props.onChangePage(event, this.props.page + 1);
-    };
-
-    handleLastPageButtonClick = event => {
-        this.props.onChangePage(
-            event,
-            Math.max(0, Math.ceil(this.props.count / this.props.rowsPerPage) - 1)
-        );
-    };
-
-    render() {
-        const { classes, count, page, rowsPerPage, theme } = this.props;
-
-        return (
-            <div className={classes.root}>
-                <IconButton
-                    onClick={this.handleFirstPageButtonClick}
-                    disabled={page === 0}
-                    aria-label="First Page"
-                >
-                    {theme.direction === 'rtl' ? <LastPageIcon /> : <FirstPageIcon />}
-                </IconButton>
-                <IconButton
-                    onClick={this.handleBackButtonClick}
-                    disabled={page === 0}
-                    aria-label="Previous Page"
-                >
-                    {theme.direction === 'rtl' ? <KeyboardArrowRight /> : <KeyboardArrowLeft />}
-                </IconButton>
-                <IconButton
-                    onClick={this.handleNextButtonClick}
-                    disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                    aria-label="Next Page"
-                >
-                    {theme.direction === 'rtl' ? <KeyboardArrowLeft /> : <KeyboardArrowRight />}
-                </IconButton>
-                <IconButton
-                    onClick={this.handleLastPageButtonClick}
-                    disabled={page >= Math.ceil(count / rowsPerPage) - 1}
-                    aria-label="Last Page"
-                >
-                    {theme.direction === 'rtl' ? <FirstPageIcon /> : <LastPageIcon />}
-                </IconButton>
-            </div>
-        );
-    }
-}
-TablePaginationActions.propTypes = {
-    classes: PropTypes.shape({}).isRequired,
-    count: PropTypes.number.isRequired,
-    onChangePage: PropTypes.func.isRequired,
-    page: PropTypes.number.isRequired,
-    rowsPerPage: PropTypes.number.isRequired,
-    theme: PropTypes.shape({}).isRequired
-};
-
 const TablePaginationActionsWrapped = withStyles(actionsStyles, { withTheme: true })(
     TablePaginationActions
 );
 
 function SalesPackages({ items, loading, rowsPerPage, page }) {
     const [rowOpen, setRowOpen] = useState();
-    const [pageLocal, setPage] = useState({});
+    const [localPage, setPage] = useState(page);
+    const [localRowsPerPage, setRowsPerpage] = useState(rowsPerPage);
 
     const handleRowOnClick = salesPackageId =>
         rowOpen === salesPackageId ? setRowOpen(null) : setRowOpen(salesPackageId);
@@ -117,7 +49,8 @@ function SalesPackages({ items, loading, rowsPerPage, page }) {
     };
 
     const handleChangeRowsPerPage = event => {
-        setPage({ page: 0, rowsPerPage: event.target.value });
+        setPage(0);
+        setRowsPerpage({ rowsPerPage: event.target.value });
     };
 
     return (
@@ -181,8 +114,8 @@ function SalesPackages({ items, loading, rowsPerPage, page }) {
                                     rowsPerPageOptions={[5, 10, 25]}
                                     colSpan={3}
                                     count={items.length}
-                                    rowsPerPage={rowsPerPage}
-                                    page={pageLocal}
+                                    rowsPerPage={localRowsPerPage}
+                                    page={localPage}
                                     SelectProps={{
                                         native: true
                                     }}
